Move payment method name to gateway converter to Behat-related services
