+++ src/Sylius/Behat/Resources/config/services/contexts/setup.xml
@@ -109,7 +109,7 @@
             <argument type="service" id="__symfony__.sylius.repository.payment_method" />
             <argument type="service" id="__symfony__.sylius.factory.payment_method" />
             <argument type="service" id="__symfony__.sylius.factory.payment_method_translation" />
-            <argument type="service" id="__symfony__.sylius.payment_method_name_to_gateway_converter" />
+            <argument type="service" id="sylius.behat.payment_method_name_to_gateway_converter" />
             <argument type="service" id="__symfony__.doctrine.orm.entity_manager" />
             <tag name="fob.context_service" />
         </service>
