+++ src/Sylius/Behat/Service/PaymentMethodNameToGatewayConverter.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CoreBundle\Test\Services;
+namespace Sylius\Behat\Service;
 
 use Payum\Core\GatewayInterface;
 use Payum\Core\Registry\RegistryInterface;
