+++ src/Sylius/Behat/Service/PaymentMethodNameToGatewayConverterInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CoreBundle\Test\Services;
+namespace Sylius\Behat\Service;
 
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
