+++ src/Sylius/Bundle/CoreBundle/Resources/config/test_services.xml
@@ -18,9 +18,6 @@
     <services>
         <!--Paypal-->
         <service id="sylius.invoice_number_generator" class="Sylius\Component\Core\Payment\RandomInvoiceNumberGenerator" />
-        <service id="sylius.payment_method_name_to_gateway_converter" class="Sylius\Bundle\CoreBundle\Test\Services\PaymentMethodNameToGatewayConverter">
-            <argument type="service" id="payum" />
-        </service>
 
         <service id="sylius.behat.email_checker" class="Sylius\Component\Core\Test\Services\EmailChecker">
             <argument>%kernel.cache_dir%/spool</argument>
